Fix: refactor `Float128.cs` into multiple files
